moving apiServices to Activity Scope (hope it is good idea)
