+++ app/src/main/java/com/example/mrr/fortnitetracker/dagger/components/FortniteApplicationComponent.java
@@ -3,8 +3,7 @@ package com.example.mrr.fortnitetracker.dagger.components;
 import android.app.Application;
 
 import com.example.mrr.fortnitetracker.FortniteTrackerApplication;
-import com.example.mrr.fortnitetracker.dagger.modules.FortniteApiServiceModule;
-import com.example.mrr.fortnitetracker.dagger.modules.TwitterApiModule;
+import com.example.mrr.fortnitetracker.dagger.modules.NetworkModule;
 import com.example.mrr.fortnitetracker.dagger.scopes.FortniteApplicationScope;
 
 import dagger.BindsInstance;
@@ -12,7 +11,7 @@ import dagger.Component;
 import dagger.android.AndroidInjectionModule;
 
 @FortniteApplicationScope
-@Component(modules = {AndroidInjectionModule.class, FortniteApiServiceModule.class, TwitterApiModule.class, ActivityBuilder.class })
+@Component(modules = {AndroidInjectionModule.class, NetworkModule.class, ActivityBuilder.class })
 public interface FortniteApplicationComponent {
 
     @Component.Builder
