--- app/src/main/java/com/example/mrr/fortnitetracker/dagger/modules/FortniteApiServiceModule.java
@@ -1,32 +1,0 @@
-package com.example.mrr.fortnitetracker.dagger.modules;
-
-import com.example.mrr.fortnitetracker.Network.FortniteApiService;
-import com.example.mrr.fortnitetracker.dagger.scopes.FortniteApplicationScope;
-import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory;
-
-import dagger.Module;
-import dagger.Provides;
-import okhttp3.OkHttpClient;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
-
-@Module(includes = NetworkModule.class)
-public class FortniteApiServiceModule {
-
-    @Provides
-    @FortniteApplicationScope
-    public FortniteApiService fortniteApiService(Retrofit retrofit) {
-        return retrofit.create(FortniteApiService.class);
-    }
-
-    @Provides
-    @FortniteApplicationScope
-    public Retrofit retrofit(OkHttpClient okHttpClient) {
-        return new Retrofit.Builder()
-                .addConverterFactory(GsonConverterFactory.create())
-                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-                .client(okHttpClient)
-                .baseUrl("https://api.fortnitetracker.com/v1/profile/")
-                .build();
-    }
-}
